requirements.txt
mcp>=0.1.7
pytz
python-dotenv
fastmcp>=0.2.0 
pandas>=1.3.0
jieba>=0.42.1
numpy>=1.20.0
wordcloud>=1.8.1
matplotlib>=3.4.0
scikit-learn>=0.24.0
requests>=2.28.0 